feature[TW29557]: Align Certificate Management Page URL aligned With Notice And Clear Stale Certificates (#591)
* feature[TW29557]: Align Certificate Management Page URL aligned With Notice And Clear Stale Certificates

* feature[TW29557]: Make tests dry

* feature[TW29557]: Update style

* feature[TW29557]: Import change.

* feature[TW29557]: Test Update

* feature[TW29557]: Use isInvalid

+++ plugins/org.eclipse.osee.orcs.core/src/org/eclipse/osee/orcs/core/internal/EmailCertificateServiceImpl.java
@@ -87,6 +87,11 @@ public class EmailCertificateServiceImpl implements EmailCertificateService {
    public void deletePublicCertificateForCurrentUser() {
       UserToken user = orcsApi.userService().getUser();
 
+      String existing = getPublicCertificateForCurrentUser();
+      if (Strings.isInvalid(existing)) {
+         return; // Nothing to delete
+      }
+
       TransactionBuilder tx = orcsApi.getTransactionFactory().createTransaction(CoreBranches.COMMON,
          "Delete email public certificate associated with user.");
       tx.deleteSoleAttribute(user.getArtifactId(), CoreAttributeTypes.EmailPublicCertificate);
